chore: add delegate stake example
Abstract common examples logic

Refs #202
